refactor: data-i18n キーの抽出を共有モジュールへ切り出す
guide-cases.test.ts の collectHtmlKeys は messages.test.ts の逐語コピー
で、手動同期に頼っていた。属性の書き方が増えたときに片方だけ古い抽出
のまま残ると、キーを取りこぼしてもレシピとケースの対応検査は緑のまま
通るため、劣化に気づけない。実装を html-keys.ts に 1 本化する。
